% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_schellens_et_al_2015_sup_1.R
\name{download_schellens_et_al_2015_sup_1}
\alias{download_schellens_et_al_2015_sup_1}
\title{Downloads the XLSX file by Schellens et al., 2015}
\usage{
download_schellens_et_al_2015_sup_1(
  url = "http://richelbilderbeek.nl/schellens_et_al_2015_s_1.xlsx",
  xlsx_filename = tempfile(fileext = ".xlsx"),
  verbose = FALSE
)
}
\arguments{
\item{url}{the download URL. Note that the original URL is
\url{https://doi.org/10.1371/journal.pone.0136417.s005},
which redirects to an unknown (and hence unusable)
actual download URL}

\item{xlsx_filename}{the XLSX filename}

\item{verbose}{set to TRUE for more output}
}
\value{
the XLSX filename of the downloaded file
}
\description{
Downloads the XLSX file by Schellens et al., 2015
}
\seealso{
use \link{get_schellens_et_al_2015_sup_1} to
  read the table as a \link{tibble}[tibble]
}
