% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_tmhmm_topology.R
\name{download_tmhmm_topology}
\alias{download_tmhmm_topology}
\title{Downloads the \code{.tmhmm} file by Bianchi et al., 2017}
\usage{
download_tmhmm_topology(
  url = "http://richelbilderbeek.nl/UP000005640_9606.tmhmm",
  tmhmm_filename = tempfile(fileext = ".tmhmm"),
  verbose = FALSE
)
}
\arguments{
\item{url}{the download URL,
which is \url{http://richelbilderbeek.nl/UP000005640_9606.tmhmm}}

\item{tmhmm_filename}{the TMHMM filename}

\item{verbose}{set to TRUE for more output}
}
\value{
the TMHMM filename of the downloaded file
}
\description{
Downloads the \code{.tmhmm} file by Bianchi et al., 2017
}
\seealso{
use \link{get_tmhmm_topology} to
  read the table as a \link[tibble]{tibble}
}
