% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proteome.R
\name{get_proteome}
\alias{get_proteome}
\title{Get the proteome from
Bianchi et al., 2017.
Assumes the file is already downloaded.
Use \link{download_proteome}
to download the file.}
\usage{
get_proteome(fasta_gz_filename, fasta_filename = tempfile(fileext = ".fasta"))
}
\arguments{
\item{fasta_gz_filename}{the \code{.fasta.gz} filename}

\item{fasta_filename}{name of a temporary FASTA file}
}
\value{
a \link{tibble}[tibble], with columns named
\code{name} and \code{sequences}
}
\description{
Get the proteome from
Bianchi et al., 2017.
Assumes the file is already downloaded.
Use \link{download_proteome}
to download the file.
}
