% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schellens_et_al_2015_sup_1.R
\name{get_schellens_et_al_2015_sup_1}
\alias{get_schellens_et_al_2015_sup_1}
\title{Get the table of Supplementary Material 1 from
Schellens et al., 2015.
Assumes the file is already downloaded.
Use \link{download_schellens_et_al_2015_sup_1}
to download the file.}
\usage{
get_schellens_et_al_2015_sup_1(xlsx_filename)
}
\arguments{
\item{xlsx_filename}{the XLSX filename}
}
\value{
a \link{tibble}[tibble], with columns named
\code{cell_line},
\code{infected},
\code{epitope_sequence},
\code{epitope_length},
\code{copy_number},
\code{source_protein},
\code{gene},
\code{best_allele},
\code{best_allele_rank},
\code{best_allele_predscore},
\code{normalized_copyno}
}
\description{
Get the table of Supplementary Material 1 from
Schellens et al., 2015.
Assumes the file is already downloaded.
Use \link{download_schellens_et_al_2015_sup_1}
to download the file.
}
