% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tmhmm_topology.R
\name{get_tmhmm_topology}
\alias{get_tmhmm_topology}
\title{Get the topology as predicted by TMHMM from
Bianchi et al., 2017.
Assumes the file is already downloaded.
Use \link{download_tmhmm_topology}
to download the pre-calculated topology.}
\usage{
get_tmhmm_topology(tmhmm_filename)
}
\arguments{
\item{tmhmm_filename}{name of the TMHMM file}
}
\value{
a \link[tibble]{tibble}, with columns named
\code{name} and \code{sequences}
}
\description{
Get the topology as predicted by TMHMM from
Bianchi et al., 2017.
Assumes the file is already downloaded.
Use \link{download_tmhmm_topology}
to download the pre-calculated topology.
}
