% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_proteome.R
\name{download_proteome}
\alias{download_proteome}
\title{Downloads the \code{fasta.gz} file by Bianchi et al., 2017}
\usage{
download_proteome(
 
    url = "ftp://ftp.ebi.ac.uk/pub/databases/reference_proteomes/QfO/Eukaryota/UP000005640_9606.fasta.gz",
  fasta_gz_filename = tempfile(fileext = ".fasta.gz"),
  verbose = FALSE
)
}
\arguments{
\item{url}{the download URL,
which is \url{ftp://ftp.ebi.ac.uk/pub/databases/reference_proteomes/QfO/Eukaryota/UP000005640_9606.fasta.gz}}

\item{fasta_gz_filename}{the \code{fasta.gz} filename}

\item{verbose}{set to TRUE for more output}
}
\value{
the \code{fasta.gz} name of the downloaded file
}
\description{
Downloads the \code{fasta.gz} file by Bianchi et al., 2017
}
