% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schellens_et_al_2015_sup_1.R
\name{get_schellens_et_al_2015_sup_1}
\alias{get_schellens_et_al_2015_sup_1}
\title{Get the table of Supplementary Material 1 from
Schellens et al., 2015.
Will download the file if needed.}
\usage{
get_schellens_et_al_2015_sup_1(
  url = "http://richelbilderbeek.nl/schellens_et_al_2015_s_1.xlsx",
  xlsx_filename = file.path(rappdirs::user_data_dir(appname = "bianchietal2017"),
    "schellens_et_al_2015_s_1.xlsx"),
  verbose = FALSE
)
}
\arguments{
\item{xlsx_filename}{the XLSX filename}
}
\value{
a \link{tibble}[tibble], with columns named
\code{cell_line},
\code{infected},
\code{epitope_sequence},
\code{epitope_length},
\code{copy_number},
\code{source_protein},
\code{gene},
\code{best_allele},
\code{best_allele_rank},
\code{best_allele_predscore},
\code{normalized_copyno}
}
\description{
Get the table of Supplementary Material 1 from
Schellens et al., 2015.
Will download the file if needed.
}
